test: budget affected reader audits
